Add a World argument to WorldtypeEvent
This adds a World argument to WorldTypeEvent.BiomeSize and
WorldTypeEvent.InitBiomeGens, so that those events can be
targeted to specific dimensions by mods.

The argument is optional (Nullable), because the BiomeProvider is
sometimes created without access to a World, such as from the
AnvilSaveConverter when converting pre-Anvil worlds to Anvil format.